+++ pkg/clients/streams/out.go
@@ -5,7 +5,6 @@ import (
 	"os"
 
 	"github.com/docker/docker/pkg/term"
-	"github.com/sirupsen/logrus"
 )
 
 // Out is an output stream used by the DockerCli to write normal program
@@ -35,7 +34,6 @@ func (o *Out) GetTtySize() (uint, uint) {
 	}
 	ws, err := term.GetWinsize(o.fd)
 	if err != nil {
-		logrus.Debugf("Error getting size: %s", err)
 		if ws == nil {
 			return 0, 0
 		}