+++ pkg/config/parse_test.go
@@ -2,6 +2,10 @@ package config
 
 import (
 	"os"
+	"path/filepath"
+	"testing"
+
+	"github.com/stretchr/testify/assert"
 )
 
 func setup() func() {
@@ -10,6 +14,38 @@ func setup() func() {
 	return func() {
 		os.Unsetenv("SHIPYARD_CONFIG")
 	}
+}
+
+func TestRunParsesBlueprintInMarkdownFormat(t *testing.T) {
+	absoluteFolderPath, err := filepath.Abs("../../functional_tests/test_fixtures/single_container")
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	c := New()
+	err = ParseFolder(absoluteFolderPath, c)
+	assert.NoError(t, err)
+
+	assert.NotNil(t, c.Blueprint)
+
+	assert.Equal(t, "Nic Jackson", c.Blueprint.Author)
+	assert.Equal(t, "Single Container Example", c.Blueprint.Title)
+	assert.Equal(t, "single_container", c.Blueprint.Slug)
+	assert.Equal(t, []string{"http://localhost:8080"}, c.Blueprint.BrowserWindows)
+	assert.Contains(t, c.Blueprint.Intro, "# Single Container")
+}
+
+func TestRunParsesBlueprintInHCLFormat(t *testing.T) {
+	absoluteFolderPath, err := filepath.Abs("../../functional_tests/test_fixtures/single_k3s_cluster")
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	c := New()
+	err = ParseFolder(absoluteFolderPath, c)
+	assert.NoError(t, err)
+
+	assert.NotNil(t, c.Blueprint)
 }
 
 /*